third_party/WebKit/LayoutTests/inspector-protocol/timeline/timeline-layout.html

<html>
<head>
<style>
.my-class {
    min-width: 100px;
    background-color: red;
}
</style>
<script type="text/javascript" src="../../http/tests/inspector-protocol/inspector-protocol-test.js"></script>
<script type="text/javascript" src="../../http/tests/inspector-protocol/tracing-test.js"></script>
<script>

function performActions()
{
    var div = document.querySelector("#myDiv");
    div.classList.add("my-class");
    div.offsetWidth;
    return Promise.resolve();
}

function test()
{
    InspectorTest.invokeAsyncWithTracing("performActions", finish);

    function finish(devtoolsEvents)
    {
        var schedRecalc = InspectorTest.findEvent("ScheduleStyleRecalculation", "I");
        var recalcBegin = InspectorTest.findEvent("UpdateLayoutTree", "B");
        var recalcEnd = InspectorTest.findEvent("UpdateLayoutTree", "E");
        InspectorTest.log("UpdateLayoutTree frames match: " + (schedRecalc.args.data.frame === recalcBegin.args.beginData.frame));
        InspectorTest.log("UpdateLayoutTree elementCount > 0: " + (recalcEnd.args.elementCount > 0));

        var invalidate = InspectorTest.findEvent("InvalidateLayout", "I");
        var layoutBegin = InspectorTest.findEvent("Layout", "B");
        var layoutEnd = InspectorTest.findEvent("Layout", "E");

        InspectorTest.log("InvalidateLayout frames match: " + (recalcBegin.args.beginData.frame === invalidate.args.data.frame));

        var beginData = layoutBegin.args.beginData;
        InspectorTest.log("Layout frames match: " + (invalidate.args.data.frame === beginData.frame));
        InspectorTest.log("dirtyObjects > 0: " + (beginData.dirtyObjects > 0));
        InspectorTest.log("totalObjects > 0: " + (beginData.totalObjects > 0));

        var endData = layoutEnd.args.endData;
        InspectorTest.log("has rootNode id: " + (endData.rootNode > 0));
        InspectorTest.log("has root quad: " + !!endData.root);

        InspectorTest.log("SUCCESS: found all expected events.");
        InspectorTest.completeTest();
    }

}
</script>
</head>
<body onLoad="runTest();">
<div id="myDiv">DIV</div>
</body>
</html>
